local_ETL_runner/requirements.txt
boto3>=1.34.0
botocore>=1.34.0
PyYAML>=6.0.1
pydantic>=2.5.0
click>=8.1.7
rich>=13.7.0
python-dateutil>=2.8.2
tqdm>=4.66.0
colorlog>=6.8.0
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0
moto>=4.2.0
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0
